clear_cron: skip the -u option if user is empty
Red Hat based distributions use cronie (a fork of vixie-cron / isc cron)
that do not allow the use of -u for non-root users, even if it matches
the current user.

This results in the error: "must be privileged to use -u"

Call "crontab -r" without "-u user" when the user parameter is empty.

Resolves #48